final/book-trend-agent/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
asgiref==3.6.0
astor==0.8.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dill==0.3.4
dnspython==2.3.0
docker-pycreds==0.4.0
entrypoints==0.4
exceptiongroup==1.1.0
fastapi==0.68.1
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
huggingface-hub==0.11.1
idna==3.4
imageio==2.25.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
multiprocess==0.70.12.2
networkx==3.0
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.4.0
promise==2.3
protobuf==3.20.3
psutil==5.9.4
pyarrow==10.0.1
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.4
pydeck==0.8.0
pymongo==4.3.3
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
scikit-image==0.19.3
scipy==1.10.1
selenium==4.8.0
sentry-sdk==1.14.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.14.2
streamlit==1.2.0
tifffile==2023.1.23.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
transformers==4.25.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.15.0
validators==0.20.0
wandb==0.12.17
watchdog==2.2.1
wcwidth==0.2.6
webdriver-manager==3.8.5
wsproto==1.2.0
xxhash==3.2.0
yarl==1.8.2

tweepy~=4.13.0
PyMySQL~=1.0.2
SQLAlchemy~=2.0.7